move utils/extractors.py > middleware/logging/extractors.py
