+++ docs/lib/reference/middleware/logging/extractors.rst
@@ -1,0 +1,6 @@
+extractors
+==========
+
+
+.. automodule:: starlite.middleware.logging.extractors
+    :members:
