+++ docs/lib/reference/middleware/logging/index.rst
@@ -1,0 +1,7 @@
+logging
+=======
+
+.. toctree::
+
+    middleware
+    extractors
