+++ docs/lib/reference/middleware/logging/middleware.rst
@@ -1,0 +1,5 @@
+middleware
+==========
+
+.. automodule:: starlite.middleware.logging.middleware
+    :members:
