--- docs/lib/reference/utils/extractors.rst
@@ -1,5 +1,0 @@
-extractors
-==========
-
-.. automodule:: starlite.utils.extractors
-    :members:
