+++ docs/lib/reference/utils/index.rst
@@ -4,8 +4,6 @@ starlite.utils
 .. toctree::
     :maxdepth: 1
 
-    exception
-    extractors
     pagination
     predicates
     scope
