+++ starlite/middleware/logging/__init__.py
@@ -1,0 +1,3 @@
+from .middleware import LoggingMiddleware, LoggingMiddlewareConfig
+
+__all__ = ("LoggingMiddleware", "LoggingMiddlewareConfig")
