+++ tests/utils/test_extractors.py
@@ -6,10 +6,12 @@ from starlite import Request, Response
 from starlite.connection.base import empty_receive
 from starlite.datastructures import Cookie
 from starlite.enums import RequestEncodingType
+from starlite.middleware.logging.extractors import (
+    ConnectionDataExtractor,
+    ResponseDataExtractor,
+)
 from starlite.status_codes import HTTP_200_OK
 from starlite.testing import RequestFactory
-from starlite.utils import ConnectionDataExtractor
-from starlite.utils.extractors import ResponseDataExtractor
 
 factory = RequestFactory()
